// src/app/api/model/salonUpdateProfilDTO.ts
/**
 * SalonMaster API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SalonUpdateProfilDTO { 
    nom: string;
    description: string;
    adresse: string;
    latitude: number;
    longitude: number;
}